Adding support for Casper JUnit output
We are using the standard casper xunit output but it isn’t very nice
when you use the standard JUnit parser for res.
This creates much nicer representation.

JUnit gives you one ‘Suites’ but we want to use the files the tests
have come from as the ‘Suites’. So we do this using the packagenames
function to find unique packagenames on the individual suites and
create a list of them.

JUnit also gives you a system-out in every test suite. We’re just
checking for testcases inside the suites so we don’t get these anymore.

We have also restricted the length of the names of testsuties to ensure
they fit into databases. I think we may want to improve on this in the
future.

This work was done very much in conjunction with:
https://github.com/Asimk21